requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
scikit-learn==1.6.1
numpy==2.0.2
pandas==2.3.3
joblib==1.5.3
python-dotenv==1.2.1
python-json-logger==4.0.0